Converter for pre tag
For it to work correctly, I introduced a $isChildOfPreTag property in
the AbstractNodeConverter class. The PreConverter gets the NodeConverter
instances for all it's children itself, from the converter (Html2Text)
instance, and always sets that property to true. When the property is
set to true, the getNodeText() method doesn't trim the node's text.

Besides that, also moved the methods to check if some node is a block
or inline element, to the Utils class.